feature/Fixed image not resizing properly and title readjusts mb when the screen is smaller
