+++ .idx/integrations.json
@@ -1,0 +1,3 @@
+{
+  "gemini_api": {}
+}
